Added files via upload
Originally our project was meant to incorporate LittleBits to recreate the mid 1950's children's toy "The Red Raven Magic Mirror," a variant of the praxinoscope. A 16 sided mirror sits onto of a record player and reflects the images on the inside label of the record creating an animated image along with music. 

We first created the base of the box using foam board. We did this by making box joints that allowed users to easily take apart and put back together the project.